requirements.txt
Django==4.2.7
psycopg2-binary==2.9.5
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==4.1.0
openpyxl==3.1.2
openai==1.3.7